Added setup module to abstract utility code from test cases, moved code from test_parser to that for use in other test modules
